Upgrade packages, various project cleanup.
