+++ src/steps/access.ts
@@ -9,7 +9,7 @@ import {
 } from '@jupiterone/integration-sdk-core';
 
 import { createAPIClient } from '../client';
-import { IntegrationConfig } from '../types';
+import { IntegrationConfig } from '../config';
 import { ACCOUNT_ENTITY_KEY } from './account';
 
 export async function fetchUsers({
