--- src/validateInvocation.ts
@@ -1,22 +1,0 @@
-import {
-  IntegrationExecutionContext,
-  IntegrationValidationError,
-} from '@jupiterone/integration-sdk-core';
-
-import { createAPIClient } from './client';
-import { IntegrationConfig } from './types';
-
-export default async function validateInvocation(
-  context: IntegrationExecutionContext<IntegrationConfig>,
-) {
-  const { config } = context.instance;
-
-  if (!config.clientId || !config.clientSecret) {
-    throw new IntegrationValidationError(
-      'Config requires all of {clientId, clientSecret}',
-    );
-  }
-
-  const apiClient = createAPIClient(config);
-  await apiClient.verifyAuthentication();
-}
